syncAccounts: don't stop reused drivers when folders are done
The sync account engine requires both side drivers to start processing a new
account name. The foldersArchitect must consider this case to not stop the
reused drivers when the sync of all the folders is done.

In 2f2072e30b we stopped waiting for the folders to get synced. Re-introduce the
required suspend time: the sync account engine must wait before starting the
processing of the next account name.

Driver runner: code refactoring.

Signed-off-by: Nicolas Sebrecht <nicolas.s-dev@laposte.net>